Move AnimaKai to decoders package on app module
