+++ app/src/test/java/brunodles/animewatcher/decoders/AnimaKaiFactoryTest.kt
@@ -1,4 +1,4 @@
-package brunodles.animakai
+package brunodles.animewatcher.decoders
 
 import brunodles.animewatcher.explorer.Episode
 import brunodles.animewatcher.testhelper.FactoryChecker
@@ -22,7 +22,7 @@ class AnimaKaiFactoryTest {
                 image = "http://www.animekai.info/2017/10/Screenshot_33.jpg",
                 link = "https://www.animekaionline.com/tsukipro-the-animation/episodio-1",
                 //                video = "http://www.blogger.com/video-play.mp4?contentId=1e5c9ab13c70079b",
-                video = "https://www.blogger.com/video-play.mp4?contentId=b186c220e9973f58",
+                video = "http://www.blogger.com/video-play.mp4?contentId=b186c220e9973f58",
                 nextEpisodes = arrayListOf()
         )
     }
