--- decoders/animakai/build.gradle
@@ -1,1 +1,0 @@
-apply from: rootProject.file("gradle/decoder_build.gradle")
