+++ settings.gradle
@@ -2,7 +2,6 @@ include ':explorer', ':decodertesthelper'
 include ':app'
 include ':cli'
 //include ':testcacher'
-include ':decoders:animakai'
 include ':decoders:animesonlinebr'
 include ':decoders:animesorion'
 include ':decoders:animesproject'
